💫 Add animejs for dynamic letter animation on error page
Introduces animejs as a dependency to enable dynamic letter animation on
the error page, enhancing user engagement through visually appealing
effects. This is achieved by animating the individual letters of the
'404' text. The playAnimation method is used to trigger the animation on
component mount. The inclusion of animejs not only improves user
experience but also aligns with the modern web design aesthetic
promoting interactivity and engagement.

#ISSUE-123